Working concept on moving router things to /core so it's included in deployed code.
Next is to test it in the deployed environment, and eventually roll in webpack
