+++ packages/devtools/package.json
@@ -6,7 +6,6 @@
     "@babel/eslint-parser": "^7.18.9",
     "@babel/parser": "^7.25.3",
     "@babel/traverse": "^7.25.3",
-    "@friggframework/core": "^2.0.0-next.0",
     "@friggframework/test": "^2.0.0-next.0",
     "@hapi/boom": "^7.4.11",
     "@inquirer/prompts": "^5.3.8",
@@ -30,11 +29,11 @@
   "devDependencies": {
     "@friggframework/eslint-config": "^2.0.0-next.0",
     "@friggframework/prettier-config": "^2.0.0-next.0",
+    "serverless": "3.39.0",
     "serverless-dotenv-plugin": "^6.0.0",
     "serverless-offline": "^13.8.0",
     "serverless-offline-sqs": "^8.0.0",
-    "serverless-webpack": "^5.14.1",
-    "serverless": "3.39.0"
+    "serverless-webpack": "^5.14.1"
   },
   "scripts": {
     "lint:fix": "prettier --write --loglevel error . && eslint . --fix",
